Moved common functionality to AbstractPluginManager
- AbstractPluginManager defines:
  - allowOverride as on by default
  - validatePlugin() as abstract, and __construct() adds it as an
    initializer
  - get/createFromInvokable() allow passing options for requested
    plugins
- Filter\ServiceManager now extends AbstractPluginManager, and only
  needs to define invokables, aliases, and validatePlugin()
